requirements.txt
python-telegram-bot
requests
influxdb
